Split directory mod into gateways and services
