+++ nym-connect/desktop/src-tauri/src/main.rs
@@ -86,10 +86,9 @@ fn main() {
             crate::operations::connection::status::get_connection_status,
             crate::operations::connection::status::get_gateway_connection_status,
             crate::operations::connection::status::start_connection_health_check_task,
-            crate::operations::directory::get_gateway_with_low_latency,
-            crate::operations::directory::get_gateways,
-            crate::operations::directory::get_services,
-            crate::operations::directory::select_gateway_with_low_latency_from_list,
+            crate::operations::directory::gateways::get_gateways,
+            crate::operations::directory::gateways::select_gateway_with_low_latency_from_list,
+            crate::operations::directory::services::get_services,
             crate::operations::export::export_keys,
             crate::operations::window::hide_window,
             crate::operations::growth::test_and_earn::growth_tne_get_client_id,
